Simple Reflex Agent
Consider the vacuum world shown in the figure below: This particular world has just two locations: squares A and B. The vacuum agent perceives which square it is in and whether there is dirt in the square. It can choose to move left, move right, suck up the dirt, or do nothing. One very simple agent function is the following: if the current square is dirty, then suck, otherwise move to the other square. Write a simple reflex agent for the vacuum cleaner. (Hint: Agent has no initial states knowledge) If the current square is dirty, then suck; otherwise, move to the other square. 
Pseudocode to the task is as follows; 
function Reflex-Vacuum-Agent( [location,status]) returns an action if status = Dirty then return Suck else if location = A then return Right else if location = B then return Left